Improved colors and new functions (get_revealed_cubes)
Created the functions get_revealed_cubes() and get_not_revealed_cubes(), the first one gets the coordinates and the text.value of the numbers of mine around each revealed cube. The second one gets only the coordinates of the unrevealed cubes.

Created a list of 5 colors, the whole cube color is based on one of those 5 colors (randomly choose), and then each cube is randomly created with a color that blends from black to one of the 5. 
It's not definitive, it's another way to avoid shadows / lights